+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix init service
 - No css file on element generation
+- default html content in element
 
 # <= 0.1.3